adding docker support

+++ EbayScrapperV4.py
@@ -9,20 +9,28 @@ from datetime import date
 import time
 import os
 
+debug=False
 
-path = "/app" #The path you're storing the program in
-user = "" #The name of the user you are running this with
+if (debug == False):
+    path = "/app" #The path you're storing the program in
+    user = "" #The name of the user you are running this with
+else:
+    path = "/home/roosevelt/Downloads/EbayScrapper"
+    user = "roosevelt"
 
 chrome_options = Options()
 chrome_options.add_argument("--no-sandbox") # linux only
 chrome_options.add_argument("--enable-crash-reporter")
-chrome_options.add_argument("--headless")
+chrome_options.add_argument('--window-size=1200,800');
+if (debug == False):
+    chrome_options.add_argument("--headless")
 #chrome_options.add_argument('--ignore-certificate-errors')
 #chrome_options.add_argument('--allow-running-insecure-content')
+
 driver = webdriver.Chrome(options=chrome_options)
 
 #Ebay's advanced settings don't show up if the window is too small.
-driver.set_window_size(1024, 768)
+#driver.set_window_size(1920, 380)
 
 #exit()
 
@@ -50,7 +58,9 @@ InitialSearch = driver.find_element("xpath", """//*[@id="_nkw"]""")
 InitialSearch.clear()
 InitialSearch.send_keys("Graphics Card")
 InitialSearch.submit()
-time.sleep(1)
+time.sleep(2)
+
+print("browser setup successfully!")
 
 #initialsearch = driver.find_elements("xpath", '/html/body/header/table/tbody/tr/td[5]/form/table/tbody/tr/td[1]/div[1]/div/input[1]')
 
@@ -62,48 +72,65 @@ time.sleep(1)
 #f.write("boo!\n")
 #f.close()
 
-os.chdir(path + '/Details/')
-if os.path.exists(str(date.today())) == True:
-    os.remove(str(date.today()))
-f = open(str(date.today()).replace('-',''), 'w')
-f.write(str(len(inputs)*2) + "\n")
-f.write("\n")
-os.chdir(path)
+if (debug == False):
+    os.chdir(path + '/Details/')
+    if os.path.exists(str(date.today())) == True:
+        os.remove(str(date.today()))
+    f = open(str(date.today()).replace('-',''), 'w')
+    f.write(str(len(inputs)*2) + "\n")
+    f.write("\n")
+    os.chdir(path)
 
 
 FirstRun = True
 for item in inputs:
+    prices = []
     time.sleep(1)
     print("item: " + item)
     input = item
     CombinedAverages = 0
-    TestLoop = 0
+    counter = 0
 
-    while TestLoop < 1: # I don't know why there are multiple loops every search returns the same thing?
-        TestLoop = TestLoop + 1
+    while len(prices) == 0: #repeat attempt search until items are successfully grabbed
+        if (counter >= 3):
+            print("3 failed attempts, rebooting browser...")
+            driver.quit()
+            driver = webdriver.Chrome(options=chrome_options)
+            driver.get("https://www.ebay.com/")
+            time.sleep(1)
 
-        if FirstRun == True:
-            search = driver.find_element("xpath",'//*[@id="gh-ac"]') #yes this used to matter and now it doesn't, but I'm not changing it
+            #Advanced Settings button click
+            driver.find_element("xpath", '//*[@id="gh-f"]/div[2]/a').click()
+            time.sleep(1)
 
-            search.clear()
+            #Sold Listings Only button click
+            driver.find_element("xpath", """//*[@id="s0-1-17-5[1]-[2]-LH_Sold"]""").click()
+            time.sleep(1)
 
-            search.send_keys(input + " graphics card")
+            #Do a initial search to apply filter
+            InitialSearch = driver.find_element("xpath", """//*[@id="_nkw"]""")
+            InitialSearch.clear()
+            InitialSearch.send_keys("Graphics Card")
+            InitialSearch.submit()
+            time.sleep(2)
+            counter = 0
 
-            search.submit()
-            
-            FirstRun = False
-        else: 
-            search = driver.find_element("xpath",'//*[@id="gh-ac"]')
+        print("attempting to search for item...")
+        search = driver.find_element("xpath",'//*[@id="gh-ac"]')
+        time.sleep(0.125)
+        search.clear()
+        time.sleep(0.125)
+        search.send_keys(input + " graphics card")
+        time.sleep(0.125)
+        search.submit()
 
-            search.clear()
+        time.sleep(2)
+        prices = driver.find_elements(By.XPATH, ".//span[@class='s-item__price']/span")
+        #prices = driver.find_elements(By.CSS_SELECTOR, "span.s-item__price")
+        #print(prices)
+        counter = counter + 1
 
-            search.send_keys(input + " graphics card")
-
-            search.submit()
-	
-        time.sleep(5)
-        prices = driver.find_elements(By.XPATH, ".//span[@class='s-item__price']/span[@class='POSITIVE']")
-
+    if (1==1): #pov you didn't want to fix the indentation
         AddedPrices = 0
         NumberOfPrices = 0
         counter = 0
@@ -129,15 +156,13 @@ for item in inputs:
                     NumberOfPrices = NumberOfPrices + 1
             counter = counter + 1
 
-        #print("Average selling price for test " + str(TestLoop) + ": "  + str(AddedPrices/NumberOfPrices))
-
-        print("CombinedAverages:" + str(CombinedAverages))
         print("AddedPrices: " + str(AddedPrices))
         print("NumberOfPrices:" + str(NumberOfPrices))
-        CombinedAverages = CombinedAverages + AddedPrices/NumberOfPrices #for when using multiple search loops
 
-    f.write(input + '\n')
-    f.write(str(CombinedAverages/TestLoop) + '\n')
-    print("Total Average for " + str(input) + ": " + str(CombinedAverages/TestLoop))
+    if (debug==False):
+        f.write(input + '\n')
+        f.write(str(AddedPrices/NumberOfPrices) + '\n')
+    print("Total Average for " + str(input) + ": " + str(AddedPrices/NumberOfPrices))
 
 driver.quit()
+f.close()
